requirements.txt
alembic==1.9.4
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
click==8.1.3
exceptiongroup==1.1.0
flake8==6.0.0
greenlet==2.0.2
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
